+++ internal/util/testutil/equal.go
@@ -15,22 +15,27 @@
 package testutil
 
 import (
+	"bytes"
 	"fmt"
+	"math"
 	"testing"
+	"time"
 
 	"github.com/pmezard/go-difflib/difflib"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"golang.org/x/exp/slices"
 
 	"github.com/FerretDB/FerretDB/internal/fjson"
 	"github.com/FerretDB/FerretDB/internal/types"
+	"github.com/FerretDB/FerretDB/internal/util/must"
 )
 
 // AssertEqual asserts that two BSON values are equal.
 func AssertEqual[T types.Type](t testing.TB, expected, actual T) bool {
 	t.Helper()
 
-	if types.Equal(expected, actual) {
+	if Equal(expected, actual) {
 		return true
 	}
 
@@ -43,7 +48,7 @@ func AssertEqual[T types.Type](t testing.TB, expected, actual T) bool {
 func AssertNotEqual[T types.Type](t testing.TB, expected, actual T) bool {
 	t.Helper()
 
-	if !types.Equal(expected, actual) {
+	if !Equal(expected, actual) {
 		return true
 	}
 
@@ -74,4 +79,208 @@ func diffValues[T types.Type](t testing.TB, expected, actual T) (expectedS strin
 	require.NoError(t, err)
 
 	return
+}
+
+// Equal compares any BSON values.
+func Equal[T types.Type](v1, v2 T) bool {
+	return equal(v1, v2)
+}
+
+// equal compares any BSON values.
+func equal(v1, v2 any) bool {
+	switch v1 := v1.(type) {
+	case *types.Document:
+		d, ok := v2.(*types.Document)
+		if !ok {
+			return false
+		}
+		if !equalDocuments(v1, d) {
+			return false
+		}
+
+	case *types.Array:
+		a, ok := v2.(*types.Array)
+		if !ok {
+			return false
+		}
+		if !equalArrays(v1, a) {
+			return false
+		}
+
+	default:
+		if !EqualScalars(v1, v2) {
+			return false
+		}
+	}
+
+	return true
+}
+
+// equalDocuments compares BSON documents. Nils are not allowed.
+func equalDocuments(v1, v2 *types.Document) bool {
+	if v1 == nil {
+		panic("v1 is nil")
+	}
+	if v2 == nil {
+		panic("v2 is nil")
+	}
+
+	keys := v1.Keys()
+	if !slices.Equal(keys, v2.Keys()) {
+		return false
+	}
+
+	for _, k := range keys {
+		f1 := must.NotFail(v1.Get(k))
+		f2 := must.NotFail(v2.Get(k))
+		if !equal(f1, f2) {
+			return false
+		}
+	}
+
+	return true
+}
+
+// equalArrays compares BSON arrays. Nils are not allowed.
+func equalArrays(v1, v2 *types.Array) bool {
+	if v1 == nil {
+		panic("v1 is nil")
+	}
+	if v2 == nil {
+		panic("v2 is nil")
+	}
+
+	l := v1.Len()
+	if l != v2.Len() {
+		return false
+	}
+
+	for i := 0; i < l; i++ {
+		el1 := must.NotFail(v1.Get(i))
+		el2 := must.NotFail(v2.Get(i))
+		if !equal(el1, el2) {
+			return false
+		}
+	}
+
+	return true
+}
+
+// EqualScalars compares BSON scalar values in a way that is useful for tests:
+//  * float64 NaNs are equal to each other;
+//  * time.Time values are compared using Equal method.
+func EqualScalars(v1, v2 any) bool {
+	switch s1 := v1.(type) {
+	case float64:
+		s2, ok := v2.(float64)
+		if !ok {
+			s3, ok := v2.(float32)
+			if !ok {
+				return false
+			}
+			return s1 == float64(s3)
+		}
+		if math.IsNaN(s1) {
+			return math.IsNaN(s2)
+		}
+		return s1 == s2
+	case float32:
+		s2, ok := v2.(float32)
+		if !ok {
+			s3, ok := v2.(float64)
+			if !ok {
+				return false
+			}
+			return float64(s1) == s3
+		}
+		if math.IsNaN(float64(s1)) {
+			return math.IsNaN(float64(s2))
+		}
+		return s1 == s2
+
+	case string:
+		s2, ok := v2.(string)
+		if !ok {
+			return false
+		}
+		return s1 == s2
+
+	case types.Binary:
+		s2, ok := v2.(types.Binary)
+		if !ok {
+			return false
+		}
+		return s1.Subtype == s2.Subtype && bytes.Equal(s1.B, s2.B)
+
+	case types.ObjectID:
+		s2, ok := v2.(types.ObjectID)
+		if !ok {
+			return false
+		}
+		return s1 == s2
+
+	case bool:
+		s2, ok := v2.(bool)
+		if !ok {
+			return false
+		}
+		return s1 == s2
+
+	case time.Time:
+		s2, ok := v2.(time.Time)
+		if !ok {
+			return false
+		}
+		return s1.Equal(s2)
+
+	case types.NullType:
+		_, ok := v2.(types.NullType)
+		return ok
+
+	case types.Regex:
+		s2, ok := v2.(types.Regex)
+		if !ok {
+			return false
+		}
+		return s1.Pattern == s2.Pattern && s1.Options == s2.Options
+
+	case int32:
+		s2, ok := v2.(int32)
+		if !ok {
+			s3, ok := v2.(int64)
+			if !ok {
+				return false
+			}
+			return int64(s1) == s3
+		}
+		return s1 == s2
+
+	case types.Timestamp:
+		s2, ok := v2.(types.Timestamp)
+		if !ok {
+			return false
+		}
+		return s1 == s2
+
+	case int64:
+		s2, ok := v2.(int64)
+		if !ok {
+			s3, ok := v2.(int32)
+			if !ok {
+				return false
+			}
+			return s1 == int64(s3)
+		}
+		return s1 == s2
+
+	case types.CString:
+		s2, ok := v2.(types.CString)
+		if !ok {
+			return false
+		}
+		return s1 == s2
+
+	default:
+		panic(fmt.Sprintf("unhandled types %T, %T", v1, v2))
+	}
 }
